requirements.txt
torch>=2.9.0
transformers>=4.57.0
peft>=0.18.0
bitsandbytes>=0.49.0
accelerate>=1.12.0
sentence-transformers>=5.2.0
neo4j>=6.1.0
gradio>=5.23.0
python-dotenv>=1.2.0
